Implemented RAII for UndirectedEdge in DynCon
